extract getProcessDefinitionData selector
