+++ designer/client/src/components/graph/GraphWrapped.tsx
@@ -7,7 +7,8 @@ import { useForkRef } from "rooks";
 import * as NotificationActions from "../../actions/notificationActions";
 import { useEventTracking } from "../../containers/event-tracking";
 import { getProcessCategory, getSelectionState, isPristine } from "../../reducers/selectors/graph";
-import { getLoggedUser, getProcessDefinitionData } from "../../reducers/selectors/settings";
+import { getProcessDefinitionData } from "../../reducers/selectors/processDefinitionData";
+import { getLoggedUser } from "../../reducers/selectors/settings";
 import { getUserSettings } from "../../reducers/selectors/userSettings";
 import { useWindows } from "../../windowManager";
 import { Graph } from "./Graph";
