+++ designer/client/src/components/graph/node-modal/fragment-input-definition/FragmentInputDefinition.tsx
@@ -2,7 +2,7 @@
 import React, { useCallback, useMemo } from "react";
 import { useSelector } from "react-redux";
 import ProcessUtils from "../../../../common/ProcessUtils";
-import { getProcessDefinitionData } from "../../../../reducers/selectors/settings";
+import { getProcessDefinitionData } from "../../../../reducers/selectors/processDefinitionData";
 import { MapVariableProps } from "../MapVariable";
 import { NodeCommonDetailsDefinition } from "../NodeCommonDetailsDefinition";
 import { FieldsSelect } from "./FieldsSelect";
