+++ designer/client/src/components/graph/node-modal/node/FragmentContent.tsx
@@ -2,12 +2,12 @@ import React, { useCallback, useState } from "react";
 import { useSelector } from "react-redux";
 import HttpService from "../../../../http/HttpService";
 import { getProcessCounts, getStickyNotes } from "../../../../reducers/selectors/graph";
+import { getProcessDefinitionData } from "../../../../reducers/selectors/processDefinitionData";
 import { FragmentNodeType } from "../../../../types";
 import { ErrorBoundary, DialogErrorFallbackComponent } from "../../../common/error-boundary";
 import NodeUtils from "../../NodeUtils";
 import { FragmentGraphPreview } from "../../fragmentGraph";
 import { correctFetchedDetails } from "../../../../reducers/graph/correctFetchedDetails";
-import { getProcessDefinitionData } from "../../../../reducers/selectors/settings";
 import { getFragmentNodesPrefix, useModalDetailsIfNeeded } from "../../../../containers/hooks/useModalDetailsIfNeeded";
 import { useInitEffect } from "../../../../containers/hooks/useInitEffect";
 import { Scenario } from "../../../Process/types";
