+++ designer/client/src/components/toolbars/search/AdvancedSearchFilters.tsx
@@ -5,6 +5,7 @@ import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
 import { getComponentGroups } from "../../../reducers/selectors/getComponentGroups";
 import { getScenario } from "../../../reducers/selectors/graph";
+import { getProcessDefinitionData } from "../../../reducers/selectors/processDefinitionData";
 import NodeUtils from "../../graph/NodeUtils";
 import { SearchLabel } from "../../sidePanels/SearchLabel";
 import { SearchLabeledAutocomplete } from "../../sidePanels/SearchLabeledAutocomplete";
